🏷️rename: app/utils 폴더 내부 함수들을 app/lib/utils로 변경
