+++ src/lib/utils/index.ts
@@ -1,5 +1,11 @@
 export { default as calculateAge } from "./calculateAge";
+export { default as calculatePathCenter } from "./calculatePathCenter";
 export { default as calculateWalkingCalories } from "./calculateWalkingCalories";
 export { default as convertDateToYearMonthDay } from "./convertDateToYearMonthDay";
+export { default as convertFormatDate } from "./convertFormatDate";
 export { default as convertMeter } from "./convertMeter";
 export { default as convertSeconds } from "./convertSeconds";
+export { default as drawPath } from "./drawPath";
+
+export { default as parseLocationObject } from "./parseLocation";
+export { default as pathAbleCheck } from "./pathAbleCheck";
